Function task::with_current_task

pub fn with_current_task<F, R>(function: F) -> Result<R, ()>where
    F: FnOnce(&TaskRef) -> R,

Invokes the given function with a reference to the current task.

This is useful to avoid cloning a reference to the current task.

Returns an Err if the current task cannot be obtained.