Module task::scheduler

Structs

PriorityInheritanceGuard

A guard that lowers a task’s priority back to its previous value when dropped.

Traits

PriorityScheduler

A task scheduler that supports some notion of priority.

Scheduler

A task scheduler.

Functions

add_task

Adds the given task to the least busy run queue.

add_task_to

Adds the given task to the specified CPU’s run queue.

add_task_to_current

Adds the given task to the current CPU’s run queue.

busyness

Returns the busyness of the scheduler on the given CPU, in which higher values indicate a busier scheduler.

inherit_priority

Modifies the given task’s priority to be the maximum of its priority and the current task’s priority.

priority

Returns the priority of the given task.

remove_task

Removes the given task from all run queues.

remove_task_from

Removes the given task from the specified CPU’s run queue.

remove_task_from_current

Removes the given task from the current CPU’s run queue.

schedule

Yields the current CPU by selecting a new Task to run next, and then switches to that new Task.

set_policy

Sets the scheduler policy for the given CPU.

set_priority

Sets the priority of the given task.

tasks

Returns the list of tasks running on each CPU.