Struct task::JoinableTaskRef

pub struct JoinableTaskRef { /* private fields */ }

A reference to a Task that can be joined; auto-derefs into TaskRef.

This allows another task to join this task, i.e., wait for this task to finish executing, and to obtain its ExitValue thereafter.

Drop-based Behavior

The contained Task is joinable until this object is dropped. When dropped, this task will be marked as non-joinable and treated as an “orphan” task. This means that there is no way for another task to wait for it to complete or obtain its exit value. As such, this task will be auto-reaped after it exits (in order to avoid zombie tasks).

Not Clone-able

Due to the above drop-based behavior, this type does not implement Clone because it assumes there is only ever one JoinableTaskRef per task.

However, this type auto-derefs into an inner TaskRef, which can be cloned, so you can easily call .clone() on it.

Implementations

impl JoinableTaskRef

pub fn join(&self) -> Result<ExitValue, &'static str>

Busy-waits (spins in a loop) until this task has exited or has been killed.

Return

Ok containing this Task’s ExitValue once this task has exited.

This includes cases where this Task failed or was killed.

Err if there was a problem while waiting for this task to exit.

This does not include cases where this Task failed or was killed, rather only cases where the join operation itself failed.