Struct task::Task

pub struct Task {

    pub id: usize,
    pub name: String,
    pub running_on_cpu: Option<u8>,
    pub runstate: RunState,
    pub saved_sp: usize,
    pub mmi: MmiRef,
    pub kstack: Stack,
    pub pinned_core: Option<u8>,
    pub is_an_idle_task: bool,
    pub app_crate: Option<Arc<AppCrateRef>>,
    pub namespace: Arc<CrateNamespace>,
    pub kill_handler: Option<KillHandler>,
    pub env: Arc<Mutex<Environment>>,
    pub failure_cleanup_function: FailureCleanupFunction,
    pub restart_info: Option<RestartInfo>,
    // some fields omitted

}

A structure that contains contextual information for a thread of execution.

Fields
id: usize

the unique id of this Task.

name: String

the simple name of this Task

running_on_cpu: Option<u8>

Which cpu core the Task is currently running on. None if not currently running.

runstate: RunState

the runnability status of this task, basically whether it’s allowed to be scheduled in.

saved_sp: usize

the saved stack pointer value, used for task switching.

mmi: MmiRef

Memory management details: page tables, mappings, allocators, etc. This is shared among all other tasks in the same address space.

kstack: Stack

The kernel stack, which all Tasks must have in order to execute.

pinned_core: Option<u8>

Whether or not this task is pinned to a certain core. The idle tasks (like idle_task) are always pinned to their respective cores.

is_an_idle_task: bool

Whether this Task is an idle task, the task that runs by default when no other task is running. There exists one idle task per core, so this is false for most tasks.

app_crate: Option<Arc<AppCrateRef>>

For application Tasks, this is a reference to the LoadedCrate that contains the entry function for this Task.

namespace: Arc<CrateNamespace>

This Task is linked into and runs within the context of this CrateNamespace.

kill_handler: Option<KillHandler>

The function that will be called when this Task panics or fails due to a machine exception. It will be invoked before the task is cleaned up via stack unwinding. This is similar to Rust’s built-in panic hook, but is also called upon a machine exception, not just a panic.

env: Arc<Mutex<Environment>>

The environment of the task, Wrapped in an Arc & Mutex because it is shared among child and parent tasks

failure_cleanup_function: FailureCleanupFunction

The function that should be run as a last-ditch attempt to recover from this task’s failure, e.g., this can be called when unwinding itself fails. Typically, it will point to this Task’s specific instance of spawn::task_cleanup_failure(), which has generic type parameters that describe its function signature, argument type, and return type.

restart_info: Option<RestartInfo>

Stores the restartable information of the task. Some(RestartInfo) indicates that the task is restartable.

Implementations

impl Task

pub fn new(
kstack: Option<Stack>,
failure_cleanup_function: FailureCleanupFunction
) -> Result<Task, &'static str>

Creates a new Task structure and initializes it to be non-Runnable. By default, the new Task will inherit some of the same states from the currently-running Task: its Environment, MemoryManagementInfo, CrateNamespace, and app_crate reference. If needed, those states can be changed by setting them for the returned Task.

Arguments

kstack: the optional kernel Stack for this Task to use. If not provided, a kernel stack of the default size will be allocated and used.

Note

This does not run the task, schedule it in, or switch to it.

However, it requires tasking to already be set up, i.e., the current task must be known.

pub fn is_running(&self) -> bool

returns true if this Task is currently running on any cpu.

pub fn is_runnable(&self) -> bool

Returns true if this Task is Runnable, i.e., able to be scheduled in.

Note

This does NOT mean that this Task is actually currently running, just that it is able to be run.

pub fn has_exited(&self) -> bool

Returns true if this Task has been exited, i.e., if its RunState is either Exited or Reaped.

pub fn is_application(&self) -> bool

Returns true if this is an application Task. This will also return true if this task was spawned by an application task, since a task inherits the “application crate” field from its “parent” who spawned it.

pub fn is_userspace(&self) -> bool

Returns true if this is a userspaceTask. Currently userspace support is disabled, so this always returns false.

pub fn set_kill_handler(&mut self, callback: KillHandler)

Registers a function or closure that will be called if this Task panics or otherwise fails (e.g., due to a machine exception occurring). The given callback will be invoked before the task is cleaned up via stack unwinding.

pub fn take_kill_handler(&mut self) -> Option<KillHandler>

Takes ownership of this Task’s KillHandler closure/function if one exists, and returns it so it can be invoked without holding this Task’s Mutex. After invoking this, the Task’s kill_handler will be None.

pub fn get_exit_value(&self) -> Option<&ExitValue>

Returns a reference to the exit value of this Task, if its runstate is RunState::Exited. Unlike take_exit_value, this does not consume the exit value.

pub fn take_exit_value(&mut self) -> Option<ExitValue>

Takes ownership of this Task’s exit value and returns it, if and only if this Task was in the Exited runstate.

Note

After invoking this, the Task’s runstate will be Reaped, and this Task will be removed from the system task list.

pub fn task_switch(&mut self, next: &mut Task, apic_id: u8)

Switches from the current (self) to the given next Task.

No locks need to be held to call this, but interrupts (later, preemption) should be disabled.