Struct task::TaskRef

pub struct TaskRef(_);

A shareable, cloneable reference to a Task that exposes more methods for task management, and accesses the enclosed Task by locking it.

The TaskRef type is necessary because in many places across Theseus, a reference to a Task is used. For example, task lists, task spawning, task management, scheduling, etc.

Essentially a newtype wrapper around Arc<Lock<Task>> where Lock is some mutex-like locking type. Currently, Lock is a MutexIrqSafe, so it does not allow multiple readers simultaneously; that will cause deadlock.

TaskRef implements the PartialEq trait; two TaskRefs are considered equal if they point to the same underlying Task.

Methods

impl TaskRef

pub fn new(task: Task) -> TaskRef

Creates a new TaskRef that wraps the given Task.

Also establishes the TaskLocalData struct that will be used to determine the current Task on each processor core.

pub fn join(&self) -> Result<(), &'static str>

Waits until the given task has finished executing, i.e., blocks until its runstate is RunState::Exited. Returns Ok() when the given task is actually exited, and Err() if there is a problem or interruption while waiting for it to exit.

Note

You cannot call join() on the current thread, because a thread cannot wait for itself to finish running. This will result in an Err() being immediately returned.

You cannot call join() with interrupts disabled, because it will result in permanent deadlock (well, this is only true if the requested task is running on the same cpu... but good enough for now).

pub fn kill(&self, reason: KillReason) -> Result<(), &'static str>

Kills this Task (not a clean exit) without allowing it to run to completion. The provided KillReason indicates why it was killed.

** Currently this immediately kills the task without performing any unwinding cleanup. In the near future, the task will be unwound such that its resources are freed/dropped to ensure proper cleanup before the task is actually fully killed. **

Locking / Deadlock

This method obtains a writable lock on the underlying Task in order to mutate its state.

Return

Returns Ok if the exit status was successfully set to the given KillReason.

Returns Err if this Task was already exited, and does not overwrite the existing exit status.

Note

The Task will not be halted immediately -- it will finish running its current timeslice, and then never be run again.

pub fn lock(&self) -> MutexIrqSafeGuardRef<Task>

Obtains the lock on the underlying Task in a read-only, blocking fashion. This is okay because we want to allow any other part of the OS to read the details of the Task struct.

pub fn block(&self)

Blocks this Task by setting its RunState to blocked.

pub fn unblock(&self)

Unblocks this Task by setting its RunState to runnable.

pub fn set_kill_handler(&self, callback: KillHandler)

Registers a function or closure that will be called if this Task panics or otherwise fails (e.g., due to a machine exception). The given callback will be invoked before the task is cleaned up via stack unwinding.

Locking / Deadlock

Obtains a write lock on the enclosed Task in order to mutate its state.

pub fn take_kill_handler(&self) -> Option<KillHandler>

Takes ownership of this Task's KillHandler closure/function if one exists, and returns it so it can be invoked without holding this Task's Mutex. After invoking this, the Task's kill_handler will be None.

Locking / Deadlock

Obtains a write lock on the enclosed Task in order to mutate its state.

pub fn take_exit_value(&self) -> Option<ExitValue>

Takes ownership of this Task's exit value and returns it, if and only if this Task was in the Exited runstate. After invoking this, the Task's runstate will be Reaped.

Locking / Deadlock

Obtains a write lock on the enclosed Task in order to mutate its state.

pub fn set_env(&self, new_env: Arc<Mutex<Environment>>)

Sets the Environment of this Task.

pub fn get_env(&self) -> Arc<Mutex<Environment>>

Gets a reference to this task's Environment.

pub fn get_namespace(&self) -> Arc<CrateNamespace>

Gets a reference to this task's CrateNamespace.

pub fn is_restartable(&self) -> bool