Crate task

This crate contains the Task structure for supporting multithreading, and the associated code for dealing with tasks.

To create new Tasks, use the spawn crate.

Examples

How to wait for a Task to complete (using join()) and get its exit value.

taskref.join(); // taskref is the task that we're waiting on
if let Some(exit_result) = taskref.take_exit_value() {
    match exit_result {
        ExitValue::Completed(exit_value) => {
            // here: the task ran to completion successfully, so it has an exit value.
            // We should know the return type of this task, e.g., if `isize`,
            // we would need to downcast it from Any to isize.
            let val: Option<&isize> = exit_value.downcast_ref::<isize>();
            warn!("task returned exit value: {:?}", val);
        }
        ExitValue::Killed(kill_reason) => {
            // here: the task exited prematurely, e.g., it was killed for some reason.
            warn!("task was killed, reason: {:?}", kill_reason);
        }
    }
}

Structs

PanicInfoOwned
Just like core::panic::PanicInfo, but with owned String types instead of &str references.

RestartInfo
A struct holding data items needed to restart a Task.

TASKLIST
The list of all Tasks in the system.

Task
A structure that contains contextual information for a thread of execution.

TaskInner
The parts of a Task that may be modified after its creation.

TaskRef
A shareable, cloneable reference to a Task that exposes more methods for task management, and accesses the enclosed Task by locking it.

Enums

ExitValue
The list of ways that a Task can exit, including possible return values and conditions.

KillReason
The list of possible reasons that a given Task was killed prematurely.

RunState
The set of possible runstates that a task can be in, e.g., runnable, blocked, exited, etc.

Functions

bootstrap_task
Bootstrap a new task from the current thread of execution.

get_my_current_task
Returns a reference to the current task.

get_my_current_task_id
Returns the current task’s ID.

get_task
returns a shared reference to the Task specified by the given task_id

set_my_kill_handler
Sets the kill handler function for the current Task

Type Definitions

FailureCleanupFunction
The signature of a Task’s failure cleanup function.

KillHandler
The function signature of the callback that will be invoked when a given Task panics or otherwise fails, e.g., a machine exception occurs.