Struct task::ExitableTaskRef

pub struct ExitableTaskRef { /* private fields */ }

A wrapper around TaskRef that allows this task to mark itself as exited.

This is only obtainable when a task is first switched to, specifically while it is executing the spawn::task_wrapper() function (before it proceeds to running its actual entry function).

Not Clone-able

This type does not implement Clone because it assumes there is only ever one ExitableTaskRef per task.

However, this type auto-derefs into an inner TaskRef, which can be cloned, so you can easily call .clone() on it.

Implementations

impl ExitableTaskRef

pub fn mark_as_exited(
&self,
exit_value: Box<dyn Any + Send>
) -> Result<(), &'static str>

Call this function to indicate that this task has successfully ran to completion, and that it has returned the given exit_value.

This is only usable within task cleanup functions to indicate that the current task has cleanly exited.

Return

Returns Ok if the exit status was successfully set.

Returns Err if this Task was already exited; the existing exit status will not be overwritten.

Note

The Task will not be halted immediately – it will finish running its current timeslice, and then never be run again.

pub fn mark_as_killed(&self, reason: KillReason) -> Result<(), &'static str>

Call this function to indicate that this task has been cleaned up (e.g., by unwinding) and it is ready to be marked as killed, i.e., it will never run again.

If you want to kill another task, use TaskRef::kill() instead.

This is only usable within task cleanup functions (e.g., after unwinding) to indicate that the current task has crashed or failed and has been killed by the system.

Return

Returns Ok if the exit status was successfully set.

Returns Err if this Task was already exited, and does not overwrite the existing exit status.

Note

The Task will not be halted immediately – it will finish running its current timeslice, and then never be run again.

pub fn reap_if_orphaned(&self)

Reaps this task (if orphaned) by taking and dropping its exit value and removing it from the system task list.

If this task has not been orphaned, meaning it is still joinable, then this function does nothing.

pub fn post_context_switch_action(&self) -> PreemptionGuard

Perform any actions needed after a context switch.

Currently this only does two things:

Drops any data that the original previous task (before the context switch) prepared for us to drop, as specified by TaskInner::drop_after_task_switch.

Obtains the preemption guard such that preemption can be re-enabled when it is appropriate to do so.

Note: this publicly re-exports the private TaskRef::post_context_switch_action() function for use in the early spawn::task_wrapper functions, which is the only place where an ExitableTaskRef can be obtained.