Struct task::Task

pub struct Task {
    pub id: usize,
    pub name: String,
    pub mmi: MmiRef,
    pub is_an_idle_task: bool,
    pub app_crate: Option<Arc<AppCrateRef>>,
    pub namespace: Arc<CrateNamespace>,
    pub failure_cleanup_function: FailureCleanupFunction,
    // some fields omitted
}

A structure that contains contextual information for a thread of execution.

Implementation note

Only fields that do not permit interior mutability can safely be exposed as public because we allow foreign crates to directly access task struct fields.

Fields
id: usize

The unique identifier of this Task.

name: String

The simple name of this Task.

mmi: MmiRef

Memory management details: page tables, mappings, allocators, etc. This is shared among all other tasks in the same address space.

is_an_idle_task: bool

Whether this Task is an idle task, the task that runs by default when no other task is running. There exists one idle task per core, so this is false for most tasks.

app_crate: Option<Arc<AppCrateRef>>

For application Tasks, this is effectively a reference to the mod_mgmt::LoadedCrate that contains the entry function for this Task.

namespace: Arc<CrateNamespace>

This Task is linked into and runs within the context of this CrateNamespace.

failure_cleanup_function: FailureCleanupFunction

The function that should be run as a last-ditch attempt to recover from this task’s failure, e.g., this can be called when unwinding itself fails. Typically, it will point to this Task’s specific instance of spawn::task_cleanup_failure(), which has generic type parameters that describe its function signature, argument type, and return type.

Implementations

impl Task

pub fn new(
kstack: Option<Stack>,
failure_cleanup_function: FailureCleanupFunction
) -> Result<Task, &'static str>

Creates a new Task structure and initializes it to be non-Runnable. By default, the new Task will inherit some of the same states from the currently-running Task: its Environment, MemoryManagementInfo, CrateNamespace, and app_crate reference. If needed, those states can be changed by setting them for the returned Task.

Arguments

kstack: the optional kernel Stack for this Task to use. If not provided, a kernel stack of the default size will be allocated and used.

Note

This does not run the task, schedule it in, or switch to it.

However, it requires tasking to already be set up, i.e., the current task must be known.

pub fn set_env(&self, new_env: Arc<Mutex<Environment>>)

Sets the Environment of this Task.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to mutate it.

pub fn get_env(&self) -> Arc<Mutex<Environment>>

Gets a reference to this task’s Environment.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to access it.

pub fn is_running(&self) -> bool

Returns true if this Task is currently running.

pub fn running_on_cpu(&self) -> Option<u8>

Returns the APIC ID of the CPU this Task is currently running on.

pub fn pinned_core(&self) -> Option<u8>

Returns the APIC ID of the CPU this Task is pinned on, or None if it is not pinned.

pub fn runstate(&self) -> RunState

Returns the current RunState of this Task.

pub fn is_runnable(&self) -> bool

Returns true if this Task is Runnable, i.e., able to be scheduled in.

Note

This does NOT mean that this Task is actually currently running, just that it is able to be run.

pub fn get_namespace(&self) -> &Arc<CrateNamespace>

Returns the namespace in which this Task is loaded/linked into and runs within.

pub fn with_kstack<R, F>(&self, func: F) -> R where
F: FnOnce(&Stack) -> R,

Exposes read-only access to this Task’s Stack by invoking the given func with a reference to its kernel stack.

Locking / Deadlock

Obtains the lock on this Task’s inner state for the duration of func in order to access its stack. The given func must not attempt to obtain that same inner lock.

pub fn inner_mut(&mut self) -> &mut TaskInner

Returns a mutable reference to this Task’s inner state.

Note about mutability

This function requires the caller to have a mutable reference to this Task in order to protect the inner state from foreign crates accessing it through a TaskRef auto-dereferencing into a Task. This is because you can only obtain a mutable reference to a Task before you enclose it in a TaskRef wrapper type.

Locking / Deadlock

Because this function requires a mutable reference to this Task, no locks must be obtained.

pub fn with_restart_info<R, F>(&self, func: F) -> R where
F: FnOnce(Option<&RestartInfo>) -> R,

Exposes read-only access to this Task’s RestartInfo by invoking the given func with a reference to its RestartInfo.

Locking / Deadlock

Obtains the lock on this Task’s inner state for the duration of func in order to access its stack. The given func must not attempt to obtain that same inner lock.

pub fn has_exited(&self) -> bool

Returns true if this Task has been exited, i.e., if its RunState is either Exited or Reaped.

pub fn is_application(&self) -> bool

Returns true if this is an application Task. This will also return true if this task was spawned by an application task, since a task inherits the “application crate” field from its “parent” who spawned it.

pub fn is_userspace(&self) -> bool

Returns true if this is a userspace Task. Currently userspace support is disabled, so this always returns false.

pub fn is_restartable(&self) -> bool

Returns true if this Task was spawned as a restartable task.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to access it.

pub fn set_kill_handler(&self, callback: KillHandler)

Registers a function or closure that will be called if this Task panics or otherwise fails (e.g., due to a machine exception).

The given callback will be invoked before the task is cleaned up via stack unwinding.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to mutate it.

pub fn take_kill_handler(&self) -> Option<KillHandler>

Takes ownership of this Task’s KillHandler closure/function if one exists, and returns it so it can be invoked without holding this Task’s inner lock. After invoking this, the Task’s kill_handler will be None.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to mutate it.

pub fn take_exit_value(&self) -> Option<ExitValue>

Takes ownership of this Task’s exit value and returns it, if and only if this Task was in the Exited runstate.

If this Task was in the Exited runstate, after invoking this, this Task’s runstate will be set to Reaped and this Task will be removed from the system task list.

If this Task was not in the Exited runstate, nothing is done and None is returned.

Locking / Deadlock

Obtains the lock on this Task’s inner state in order to mutate it.

pub fn block(&self)

Blocks this Task by setting its runstate to RunState::Blocked.

pub fn unblock(&self)

Unblocks this Task by setting its runstate to RunState::Runnable.

pub fn task_switch(&self, next: &Task, apic_id: u8)

Switches from the current task (self) to the given next task.

Locking / Deadlock

Obtains the locks on both this Task’s inner state and the given next Task’s inner state in order to mutate them.