Struct task::TaskRef

pub struct TaskRef(_);

A shareable, cloneable reference to a Task that exposes more methods for task management and auto-derefs into an immutable &Task reference.

The TaskRef type is necessary because in many places across Theseus, a reference to a Task is used. For example, task lists, task spawning, task management, scheduling, etc.

Equality comparisons

TaskRef implements the PartialEq and Eq traits to ensure that two TaskRefs are considered equal if they point to the same underlying Task.

Implementations

impl TaskRef

pub fn new(task: Task) -> JoinableTaskRef

Creates a new TaskRef, a shareable wrapper around the given Task.

This function also initializes the given Task’s TaskLocalData struct, which will be used to determine the current Task on each CPU.

It does not add this task to the system-wide task list or any runqueues, nor does it schedule this task in.

Return

Returns a JoinableTaskRef, which derefs into the newly-created TaskRef and can be used to “join” this task (wait for it to exit) and obtain its exit value.

pub fn join(&self) -> Result<(), &'static str>

Blocks until this task has exited or has been killed.

Returns Ok() once this task has exited, and Err() if there is a problem or interruption while waiting for it to exit.

Note

You cannot call join() on the current thread, because a thread cannot wait for itself to finish running. This will result in an Err() being immediately returned.

You cannot call join() with interrupts disabled, because it will result in permanent deadlock (well, this is only true if the requested task is running on the same cpu… but good enough for now).

pub fn kill(&self, reason: KillReason) -> Result<(), &'static str>

Kills this Task (not a clean exit) without allowing it to run to completion. The provided KillReason indicates why it was killed.

** Currently this immediately kills the task without performing any unwinding cleanup. In the near future, the task will be unwound such that its resources are freed/dropped to ensure proper cleanup before the task is actually fully killed. **

Locking / Deadlock

This method obtains a writable lock on the underlying Task’s inner state.

Return

Returns Ok if the exit status was successfully set to the given KillReason.

Returns Err if this Task was already exited, and does not overwrite the existing exit status.

Note

The Task will not be halted immediately – it will finish running its current timeslice, and then never be run again.